Lazy imports
To reduce the number of third-party module to install in a venv
when creating an alias to have a dictionary CLI tool.

FTR here the function added to my `~/.zshrc`:

    dict() {
        PYTHONPATH=~/projects/ebook-reader-dict ~/projects/deploy-dir/.pyenv/versions/3.7.4/bin/python -m scripts --get-word "$@"
    }

Then I just type `dict WORD` to have a definitions in the terminal.

Well, this is only an example as a venv is needed for me.
But reducing what is imported helps keeping the start time
low and prevent from having to install too many modules.