[FR] Remove plural
The idea is to get rid of all kind of "flexion".

The dictionary will now only contain definitions of words with their singluar form. And the Kobo will work like a charm when selecting a plural form of a word: it will display the singular one automatically, thanks to the `words` file.

FTR 255,399 words are removed from the dictionary!

Fixes #56.