[FR] Improve scripts
First above all, `fr-regions.py` was stripping parenthesis using `.strip("()")`, and it
resulted in `Géorgie (pays` instead of `Géorgie (pays)` (for example).
Secondly, multiple pages are not templates pages, and they contain a slash in their name:
so they are now skipped instead of polluting the dict.

I applied same fixes to `fr-domain-templates.py` (the parenthesis one is not relevant yet, but it is
harmless to do it).