fix/feat: download smaller Wiktionary dumps
How on earth did I not see this in 5 years?

We were using files named `pages-meta-current` which, as stated in the doc[1],
contain "all pages, current revision only", while files named `pages-articles` contain
"all pages except for talk pages, current revision only". Obviously, we do not care about talks.

And the difference is huge! For instance with the English dump, it lowers the downloaded file size
from 40.7 GB to 22.3 GB.

[1] https://meta.wikimedia.org/wiki/Data_dumps/What's_available_for_download#XML_files