fix: full Japanese support on Kobo
The only "strange" behavior is that is WORD is not found in the dictionary, then
the Kobo mechanism will look for WOR, then WO, and W finally. So, the displayed/found
definition could be completely irrelevant to the selected/searched original term.

But this is a Kobo-specific case, and only they could do something about it. It's just a FIY.