fix:(dictfile, and co.): add back a line space between pronunciations/genders and the POS
Regression introduced in 792abfd5207fc019129374d748c8efaaa180ce9f where
the start of definitions was a list tag (`<li>`) which is seen as a "block display"
while the new code was using a strong tag (`<b>`) which is seens as "inline display".

Fixing by using a paragraph (`<p>`) since it is also the standard/recommended way.

I also tried to use a line break (`<br/>`) before the `<b>`, but the output was not satisfying:
the first line break between the pronunciations/genders and the POS was too small.

Fixes #2477.